feat: LLM MAD assetConfiguration
