+++ apps/ledger-live-mobile/src/newArch/features/ModularDrawer/screens/AssetSelection/modules/createAssetConfigurationHook.ts
@@ -1,0 +1,62 @@
+import { EnhancedModularDrawerConfiguration } from "@ledgerhq/live-common/wallet-api/ModularDrawer/types";
+import { useRightBalanceModule } from "./useRightBalanceModule";
+import { AssetType } from "@ledgerhq/native-ui/pre-ldls/index";
+import { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
+import { composeHooks } from "@ledgerhq/live-common/utils/composeHooks";
+import { CurrenciesByProviderId } from "@ledgerhq/live-common/deposit/type";
+
+type Props = {
+  assetsConfiguration: EnhancedModularDrawerConfiguration["assets"];
+  currenciesByProvider?: CurrenciesByProviderId[];
+};
+
+type CreateAssetConfigurationHook = (
+  props: Props,
+) => (assets: CryptoOrTokenCurrency[]) => (CryptoOrTokenCurrency & AssetType)[];
+
+const getRightElement = (rightElement: string) => {
+  switch (rightElement) {
+    case "balance":
+      return useRightBalanceModule;
+    case "marketTrend":
+    case "undefined":
+    default:
+      return undefined;
+  }
+};
+
+const getLeftElement = (leftElement: string) => {
+  switch (leftElement) {
+    case "apy":
+    case "priceVariation":
+    case "undefined":
+    default:
+      return undefined;
+  }
+};
+
+const createAssetConfigurationHook: CreateAssetConfigurationHook = ({
+  assetsConfiguration,
+  currenciesByProvider,
+}) => {
+  const { rightElement = "undefined", leftElement = "undefined" } = assetsConfiguration ?? {};
+
+  const rightHook = getRightElement(rightElement);
+  const leftHook = getLeftElement(leftElement);
+
+  const hooks = [rightHook, leftHook].filter(Boolean) as Array<
+    (
+      assets: CryptoOrTokenCurrency[],
+      currenciesByProvider?: CurrenciesByProviderId[],
+    ) => AssetType[]
+  >;
+
+  return (assets: CryptoOrTokenCurrency[]) => {
+    const composedHook = composeHooks<CryptoOrTokenCurrency, AssetType>(
+      ...hooks.map(hook => (assets: CryptoOrTokenCurrency[]) => hook(assets, currenciesByProvider)),
+    );
+    return composedHook(assets);
+  };
+};
+
+export default createAssetConfigurationHook;
