+++ libs/ledger-live-common/.unimportedrc.json
@@ -402,6 +402,13 @@
     "src/modularDrawer/utils/haveOneCommonProvider.ts",
     "src/modularDrawer/utils/index.ts",
     "src/modularDrawer/utils/getBalanceAndFiatValue.ts",
+    "src/modularDrawer/utils/getBalanceAndFiatValueByAssets.ts",
+    "src/modularDrawer/utils/groupAccountsByAsset.ts",
+   "src/modularDrawer/utils/parseToBigNumber.ts",
+    "src/modularDrawer/utils/type.ts",
+    "src/modularDrawer/utils/calculateProviderTotal.ts",
+    "src/modularDrawer/utils/counterValueFormatter.ts",
+    "src/modularDrawer/utils/getProviderCurrency.ts",
     "src/notifications/AnnouncementProvider/helpers.ts",
     "src/reactNativeSvg.d.ts",
     "src/rxjs/operators/throwIf.ts",
