+++ libs/ledger-live-common/package.json
@@ -254,7 +254,8 @@
     "winston": "^3.4.0",
     "xstate": "^5.19.2",
     "yargs": "^17.0.0",
-    "zod": "^3.22.4"
+    "zod": "^3.22.4",
+    "i18next": "20.6.1"
   },
   "devDependencies": {
     "@ledgerhq/device-react": "workspace:^",
