+++ libs/ledger-live-common/src/modularDrawer/utils/getBalanceAndFiatValueByAssets.ts
@@ -1,13 +1,10 @@
 import { formatCurrencyUnit } from "@ledgerhq/coin-framework/lib-es/currencies/formatCurrencyUnit";
 import type { CounterValuesState } from "@ledgerhq/live-countervalues/types";
 import type { CryptoOrTokenCurrency, Currency } from "@ledgerhq/types-cryptoassets";
-import type { AccountLike } from "@ledgerhq/types-live";
-import {
-  groupAccountsByAsset,
-  type GroupedAccount,
-} from "@ledgerhq/live-common/modularDrawer/utils/groupAccountsByAsset";
-import { counterValueFormatter } from "@ledgerhq/live-common/market/utils/countervalueFormatter";
+import { groupAccountsByAsset, type GroupedAccount } from "./groupAccountsByAsset";
 import { AssetType } from "./type";
+import { AccountLike } from "@ledgerhq/types-live";
+import { counterValueFormatter } from "./counterValueFormatter";
 
 interface ExtendedAssetType extends AssetType {
   balance?: string;
@@ -65,6 +62,8 @@ const formatAssetDetails = (
     currency: targetCurrency.ticker,
     value: assetGroup.totalFiatValue.toNumber(),
     locale: userLocale,
+    allowZeroValue: true,
+    discreetMode: isDiscreetMode,
   });
 
   return {
